Add medical device questions branch
This adds four questions that a user will see if they select
Yes in response to the question Do you have medical equipment to offer?

There is additional logic on the product details controller,
since a user can enter multiple products if they click Yes
in response to the question Do you have additional products to offer?

We need to be able to edit these products at the end, so they
are assigned IDs which will enable users to modify particular
products.

There is no limit on the number of products a user can add.
If we move to using a db-backed store we should set a limit.

This encompases questions 1.1 - 1.4, and should link to question
2.