Use fragment caching for views
This is an attempt to improve on an already quite fast response time;
feel free to close if this is unnecessary / confusing, since the gains
will be pretty small.

Since we can't cache pages due to sessions, we can instead cache
'fragments' of the views. So if two users view the same thing,
e.g. a page without any user input, we can serve a fragment from
the cache.

https://guides.rubyonrails.org/caching_with_rails.html#fragment-caching

Since we can't cache pages based on the URL, we can instead make
use of Rails caching.