Don't show questions the user has skipped
We need to know which questions could possibly skipped by the user.
A list was chosen because it was the easiest way to get the
functionality working in the first instance, how ever it has the burden
of constantly needing to be updated.

The questions of how to store this routing logic needs a bigger
discussion as it could mean significant changes to how routing is
handled everywhere and this didn't feel like the appropriate place to
make those decisions.

Also, deciding which questions need to go in this list is not straight
forward. It's not just a case of "dependent" questions belonging in this
list, as at least dependent question (`additional_products`) needs to be
excluded so that the if the user has a chance to add products from the
check answers page, even if they initially answered "no" to products.