Test that expected behaviour hasn't changed
Check that all questions are displayed when the skippable questions have
a value.

N.B.
Session is some kind of rails magic and isn't a real hash ({}) even
though it pretends to be one, so we have to use merge! rather than
directly assigning a value

Also, the answers themselves don't matter for the tests, just that they
exist, so they've been DRY-ed up.