Add Google Analytics opt-in
Users won't be tracked by Google Analytics unless
they opt-in to the tracking.

This uses the cookie banner component, and copies
the approach (cookie settings page) from GOV.UK
for this service.

If a user opts in to usage tracking, we'll load
Google Analytics so we can see how users move
through the form in aggregate, in order
to improve the application.

https://trello.com/c/TObzdyej/1-add-google-analytics-tracking-on-business-volunteering-form