Use `mini_racer` in preference to `therubyracer`
The gem 'therubyracer' is dependent on an old version of libv8 with
multiple security vulnerabilities and doesn't build easily on the latest
version of macOS. Rails replaced it with the 'mini_racer' gem in the PR
rails/rails PR 29285.

Commit message shamelessly copied from https://github.com/alphagov/govuk-coronavirus-vulnerable-people-form/pull/197.

I also hit build errors with `therubyracer` on macOS Catalina while
trying to debug a flaky test, and we should be consistent with all of
these apps.