Add validation for product quantity
Checks that the value is a number. To do this we need to strip any
commas from the value.
The check against `Integer` also catches any `nil` values so
`product_quantity` can be removed from the missing fields check.

The tests fails if equipment_type is not given, so it's being added here
to make sure that it's the `product_quantity` that's actually being
tested.

Also adds a test for when product_quantity is nil. This was being
covered by the test for `REQUIRED_FIELDS` but product_quantity was
removed from that list to stop the error being displayed to the user
twice.