Add validation for product cost
Checks that the value is a number. To do this we need to strip any
commas from the value.
The check against `Float` also catches any `nil` values so
`product_cost` can be removed from the missing fields check.

At the moment it accepts any integer or decimal value, and does not
limit the number of decimal places.

The test fails if equipment_type is not given, so it's being added here
to make sure that it's the `product_quantity` that's actually being
tested.