Update pipeline yml to run quality checks
Moves around quite a bit in pipeline YML, aims are:
- Give resources shorter, more generic names. It is clear from within a
pipeline which service we are in, and long names make for things
bunching up on the UI, so generally simplify everything down to just
`git-master` or simple descriptions of what the resource actually is.

- Create a run quality check task, this should be able to take a
concourse test docker image with all the system level dependencies we
need (Redis/Postgres) and then provide all the commands needed to set up
and run the test suite + linting.

- Rearrange the triggers in the pipeline so quality checks run every
time we get a change to the master branch. These quality checks will
then be a dependency for deployments to staging and production. If they
fail, then we don't go any further.