Use the same reference ID on every submit
This is to go some ways to preventing users submitting their form response twice.

We validate that a form response is unique, so if we use the same reference ID
for each submit then we won't see duplicate form responses as a result of a user
e.g. double clicking the submit button (which requires a separate frontend fix).

The intent here is to set the reference number when a user views the page,
ensuring that the value of the reference number doesn't change when they submit
the form.

https://trello.com/c/cQxG2uZz/284