Log error if schema validation fails
Until we're sure that the schema 100% correct, we don't want to stop
users from being able to submit the form. However we still need to know
what part of the validation failed, so the validation failures are being
sent to Sentry.

Creates a custom error so that they can be easily grouped in Sentry.

The session is being converted to a hash to be validated because even
though sessions behave like a hash, but they aren't really, Rails just
pretends it's one. So if you try to validate the session against the
schema, you can get some odd results.

It seems that the data structure of the session is actually an array of
arrays,

i.e.

```
[["question", "answer"]]
```

But when the session is stored in the database, it is actually storing
it as a hash.