Update routing from expert-advice-type page
We don't always want to ask the construction_services and it_services
questions.

If the user has completed the form, but then changes their answer to the
expert_advice_type question, we only want to ask the construction
services and IT services questions if they've been newly selected,
and not force the users to navigate to the them if they've already been
answered.

The expectation is that if the user wants to change their answers to
either construction or IT services questions, they'll click on the
change links on the check answers page for those questions.