Update routing from construction_services page
We don't always want to ask the it_services question.

If the user has completed the form, but then changes their answer to the
expert_advice_type question, we only want to ask the IT services
question if it's been newly selected, and not force the users to
navigate to it if it's already been answered.

The expectation is that if the user wants to change their answers to
IT services question, they'll click on the change link on the check
answers page for that question.