Remove unecessary data from FormResponse
This removes a few unrequired fields from the FormResponse data.

It also saves the data as a hash rather than as an array. I've
added extra tests to ensure that we export data to the data team
in the same way.

Before we merge this branch we'll need to get approval from
Cabinet Office for the change.