Remove extra "How much charge" question on results
Remove extra "How much charge" question being displayed on the results
page.

The check your answers page is populated by looping through all of the
questions listed under "coronavirus_form.questions" in the I18n locale
file.

"How much charge" is a common question that is asked on multiple pages,
so it has its own reusable entry in the locale file. Unfortunately the
entry for it was also nested under "coronavirus_form.questions" meaning
that it was being automatically picked by the CheckAnswersHelper. To fix
this the entry has no been moved up level and the views and tests have
been updated to match.