Add is UK registered field
CCS can only process offers from UK companies who have a company number, so
we don't want UK companies who don't have a company number completing the
form.

Non-UK companies can still complete the form without a company number.

Add a new mandatory field to hold if the company is UK registered or not. If
yes, company number should be shown and user required to enter a valid
company number.  If no, the company number is not shown and not required.

Suitable error messages are given to the user as feedback (if appropriate).