Redact query string params
We need to test that the product_id and reference_number are redacted and sent in the correct format.

Currently, if the parameter names changed or the code used to generate the IDs changed, then there would be no test in place to test that redacting the params for GA had broken as a result.

If the product_id and reference_number patterns are not matched and redacted, then the params are dropped, so we don't risk sending PII data. The reason that we kept redacted data at all, is that the PAs advise that it is better to have [redacted-number] because it shows that there is missing data and what the redacted data is.

This change adds testing.

Co-authored-by: Ian James <ian.james@digital.cabinet-office.gov.uk>