Fixes double pageview when cookies are accepted
Two pageviews were being fired when a user accepted cookies using the cookie
banner. This is not a good thing, as it artificially inflates views on any
page where a user accepts cookies.

The cookie banner was being initalised twice, with two sets of listeners set
up to fire pageviews attached to the banner. This pull request removes one of
those, so the pageviews now only fire once when a user has accepted cookies.