Remove existing logic
The redesign requires some fundamental changes to the way that the 
check_your_answers page is currently being rendered.  

As a result, it is considered easier to remove all of the existing logic
within the check_answers_helper, and the render items loop on the page.
Then build the page back up to the desired design.

The overall architecture will have methods that supply the question 
answers from the session and supply this to a section partial for 
display.  One partial per section.  

Each partial is responsible for displaying the question answer and 
generating the 'Change' link HREF. 

Some partials have details that do not need to be displayed if the 
answer to the question is negative - usually 'No'.  So, we will need to
test for a positive answer and then display the detail part.  Negative
responses will not display the details section responses or their 
'Change' link.

This commit removes the existing code as detailed above.