Add I18N strings
The redesign requires slightly different text than is currently
available in the en.yml file.  

Instead of changing the existing text - which is appropriate to the 
pages that these elements are on and so should not be changed - a new 
section is added to specifically handle text appropriate to the 
check_your_answers page.

In order to shorten and simplify the I18N string look-up within the 
check_your_answers_helper and reduce duplication, two helper methods are
added.

This commit adds the above changes.