Add support methods
Partials will be used to render each section, a method to define the
sections (by name) and the order that they are to be rendered in is 
therefore required.  It will then be possible to loop around the 
sections list - in order - and display each section partial.

Each section can be defined as a header and detail.  The detail for each
section will be handled by the appropriate partial, but the header is
identical in structure across all sections except for the name of the 
section name itself.  A single method can handle the data required by
all section headers.

The default Rails inflection rules do not handle the pluralization of 
'foot' - mistakenly returning 'foots'.  An inflection rule is required
to handle this instance and correctly return 'feet' when appropriate.

This commit adds the above changes.