Update Check Your Answers Page
All of the section partials are now in place, so all that is left to do 
is iterate through each section (in the desired order) and display each
partial in turn.

The redesign requires a style of margin_bottom set to 8 for the title 
and each section.  A div tag that wraps each section is required so that
the appropriate CSS class can be applied to it.

This commit adds the above changes.