Temporary bug fix
The summary_list component has an additional bottom-border set and so is
incorrectly displaying two bottom borders. The redesign requires a 
single bottom border at the end of each section. 

This is a temporary fix to only display a single bottom border until the
styling can be applied to the summary_list component in the publishing 
components gem.